rename txfjs into jsProcessor
